requirements.txt
# CTA 量化系统依赖

# === 核心科学计算 ===
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
polars>=0.20.0

# === 数据存储 ===
pyarrow>=14.0.0
fastparquet>=2023.10.0
sqlalchemy>=2.0.0

# === 数据源 ===
tushare>=1.4.0
akshare>=1.12.0

# === 机器学习 ===
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.1.0

# === 深度学习 (可选) ===
torch>=2.1.0
# torchvision>=0.16.0

# === 技术指标 ===
ta-lib>=0.4.28
pandas-ta>=0.3.14b

# === 可视化 ===
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
streamlit>=1.29.0

# === 开发工具 ===
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0

# === 代码质量 ===
black>=23.12.0
isort>=5.13.0
mypy>=1.7.0
ruff>=0.1.0

# === 测试 ===
pytest>=7.4.0
pytest-cov>=4.1.0

# === 配置管理 ===
pyyaml>=6.0.0
python-dotenv>=1.0.0

# === 日志 ===
loguru>=0.7.0

# === 网络请求 ===
requests>=2.31.0
aiohttp>=3.9.0

# === 工具 ===
tqdm>=4.66.0
joblib>=1.3.0
